refactor compiler's hir into its own salsa db
+ make the tests all pass
